+++ course/format/topcoll/classes/togglelib.php
@@ -32,8 +32,6 @@
  */
 namespace format_topcoll;
 
-defined('MOODLE_INTERNAL') || die;
-
 class togglelib {
 
     // Digits used = ":;<=>?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\]^_`abcdefghijklmnopqrstuvwxy";
@@ -166,7 +164,7 @@ class togglelib {
     }
 
     /**
-     * Tells us the position of the bit within the digit for the given toggle number and digit postion in the toggles state.
+     * Tells us the position of the bit within the digit for the given toggle number and digit position in the toggles state.
      * int $togglenum - Toggle number.
      * int $togglecharpos - Digit character position.
      * returns int - Digit flag.
@@ -256,7 +254,7 @@ class togglelib {
      * @return mixed
      * @throws coding_exception
      */
-    static public function required_topcoll_param($parname) {
+    public static function required_topcoll_param($parname) {
         if (empty($parname)) {
             throw new coding_exception('required_topcoll_param() requires $parname to be specified');
         }
@@ -272,7 +270,7 @@ class togglelib {
      * @return mixed
      * @throws coding_exception
      */
-    static public function clean_topcoll_param($param) {
+    public static function clean_topcoll_param($param) {
         if (is_array($param)) {
             throw new coding_exception('clean_topcoll_param() can not process arrays.');
         } else if (is_object($param)) {