+++ course/format/topcoll/lang/en/format_topcoll.php
@@ -40,6 +40,8 @@ $string['topcollopened'] = 'Open all';
 $string['topcollclosed'] = 'Close all';
 $string['sctopenall'] = 'Open all {$a}';
 $string['sctcloseall'] = 'Close all {$a}';
+$string['toggleclose'] = 'Close';
+$string['toggleopen'] = 'Open';
 
 $string['settoggleallenabled'] = 'Toggle all enabled';
 $string['settoggleallenabled_help'] = 'Toggle all functionality enabled.';
@@ -116,11 +118,49 @@ $string['layoutstructuretopics'] = 'topics';
 $string['layoutstructureweeks'] = 'weeks';
 $string['layoutstructuredays'] = 'days';
 
-// Coursesetting - Show addtional data for modules
+// Coursesetting - Show addtional data for modules.
+$string['enableadditionalmoddata'] = 'Enable additional information';
+$string['enableadditionalmoddatadesc'] = 'This is a \'Site level\' switch to turn the activity information on or off.  It needs to be \'on\' for the related settings that operate at a course level to take effect.  As this functionality can be computationally expensive, then it is strongly suggested that you undertake full testing before using on a production system.  Note: Purges the cache caches when changed.';
+
 $string['showadditionalmoddata'] = 'Show additional information for: {$a} in the course';
 $string['showadditionalmoddata_help'] = 'Allow all users to see the activity deadline and users with grading permission to see the number of submissions on the course page for activities.';
-$string['defaultshowadditionalmoddata'] = 'Default course \'Show additional information\' if not already set';
+$string['defaultshowadditionalmoddata'] = 'Default course \'Show additional information\'';
 $string['defaultshowadditionalmoddatadesc'] = 'If an activity is set at site level (below) to show additional information then this setting states the default state of the course specific instance of it.';
+$string['resetactivitymeta'] = 'Additional module information';
+$string['resetallactivitymeta'] = 'All additional module information';
+$string['resetactivitymeta_help'] = 'Resets the additional module information to follow the site default value.';
+$string['resetallactivitymeta_help'] = 'Resets all the additional module information to follow the site default value.';
+
+$string['courseadditionalmoddatamaxstudents'] = 'Set the maximum number of students on a course that \'Show additional information\' will apply to';
+$string['courseadditionalmoddatamaxstudentsdesc'] = 'Additional information can take time to calculate, especially on large courses, so here you can set the maximum number of students that a couse can have for the functionality to show on that course.  Above that value, the \'Additional information\' will NOT be calculated or show regardless of the course settings!  A value of \'0\' means \'unlimited\'.  Note: Purges the cache caches when changed.';
+
+$string['courseadditionalmoddatastudentsinfo'] = 'Additional information status:';
+$string['courseadditionalmoddatastudentsinfounlimited'] = 'Additional information will show for the enabled activities for {$a} students.';
+$string['courseadditionalmoddatastudentsinfolimitedshow'] = 'Additional information will show for the enabled activities for {$a->students} students as the number does not exceed the maximum \'{$a->maxstudents}\' set by the administrator on the Collapsed Topics course format setting \'courseadditionalmoddatamaxstudents\'.';
+$string['courseadditionalmoddatastudentsinfolimitednoshow'] = 'Additional information will NOT show for the enabled activities for {$a->students} students as the number exceededs the maximum \'{$a->maxstudents}\' set by the administrator on the Collapsed Topics course format setting \'courseadditionalmoddatamaxstudents\'.';
+
+$string['coursesectionactivityfurtherinformation'] = 'Course page further information';
+$string['coursesectionactivityfurtherinformation_desc'] = 'Site level course page further information settings';
+$string['coursesectionactivityfurtherinformationassign'] = 'Show assignment information';
+$string['coursesectionactivityfurtherinformationassigndesc'] = 'Allow assignment information to be selected to be shown on a course.  For teachers / admins, show number of submissions.';
+$string['coursesectionactivityfurtherinformationquiz'] = 'Show quiz information';
+$string['coursesectionactivityfurtherinformationquizdesc'] = 'Allow quiz information to be selected to be shown on a course.  For teachers / admins, show number of submissions.';
+$string['coursesectionactivityfurtherinformationchoice'] = 'Show choice information';
+$string['coursesectionactivityfurtherinformationchoicedesc'] = 'Allow choice information to be selected to be shown on a course.  For teachers / admins, show number of submissions.';
+$string['coursesectionactivityfurtherinformationfeedback'] = 'Show feedback information';
+$string['coursesectionactivityfurtherinformationfeedbackdesc'] = 'Allow feedback information to be selected to be shown on a course.  For teachers / admins, show number of submissions.';
+$string['coursesectionactivityfurtherinformationforum'] = 'Show forum information';
+$string['coursesectionactivityfurtherinformationforumdesc'] = 'Allow forum information to be selected to be shown on a course.  For teachers / admins, show number of contributions when whole forum grading is on.';
+$string['coursesectionactivityfurtherinformationlesson'] = 'Show lesson information';
+$string['coursesectionactivityfurtherinformationlessondesc'] = 'Allow lesson information to be selected to be shown on a course.  For teachers / admins, show number of submissions.';
+$string['coursesectionactivityfurtherinformationdata'] = 'Show database information';
+$string['coursesectionactivityfurtherinformationdatadesc'] = 'Allow data information to be selected to be shown on a course.  For teachers / admins, show number of submissions.';
+
+$string['cachedef_activitystudentrolescache'] = 'Caches the student roles.';
+$string['cachedef_activitymodulecountcache'] = 'Caches the number of students who can access a given module on a given course.';
+$string['cachedef_activitystudentscache'] = 'Caches the ids of the students on a given course.';
+$string['cachedef_activityusercreatedcache'] = 'Caches the ids of the new users on a given course.';
+$string['cannotgetactivitycacheslock'] = 'Cannot get activity caches lock for course id {$a}.';
 
 // Colour enhancement - Moodle Tracker CONTRIB-3529.
 $string['setcolour'] = 'Colour';
@@ -134,13 +174,16 @@ $string['resetallcolour'] = 'Colours';
 
 // Columns enhancement.
 $string['setlayoutcolumns'] = 'Columns';
+$string['setlayoutcolumns_help'] = 'How many columns to use.';
 $string['one'] = 'One';
 $string['two'] = 'Two';
 $string['three'] = 'Three';
 $string['four'] = 'Four';
 $string['setlayoutcolumnorientation'] = 'Column orientation';
-$string['columnvertical'] = 'Vertical';
+$string['setlayoutcolumnorientation_help'] = 'Dynamic - Number sections per \'row\' adjust to window size, \'Column\' setting not currently used.<br>Horizontal - Sections go left to right.<br>Vertical - Sections go top to bottom.';
+$string['columndynamic'] = 'Dynamic';
 $string['columnhorizontal'] = 'Horizontal';
+$string['columnvertical'] = 'Vertical';
 
 // MDL-34917 - implemented in M2.5 but needs to be here to support M2.4- versions.
 $string['maincoursepage'] = 'Main course page';
@@ -160,8 +203,6 @@ $string['settogglebackgroundhovercolour_help'] = 'Sets the background colour of 
 $string['resetcolour_help'] = 'Resets the colours and opacities to follow the site default value.';
 $string['resetallcolour_help'] = 'Resets the colours and opacities to follow the site default value.';
 // Columns enhancement.
-$string['setlayoutcolumns_help'] = 'How many columns to use.';
-$string['setlayoutcolumnorientation_help'] = 'Vertical - Sections go top to bottom.<br />Horizontal - Sections go left to right.';
 
 // Moodle 2.4 Course format refactoring - MDL-35218.
 $string['numbersections'] = 'Number of sections';
@@ -177,7 +218,7 @@ $string['right'] = 'Right';
 $string['resettogglealignment'] = 'Toggle alignment';
 $string['resetalltogglealignment'] = 'Toggle alignments';
 $string['resettogglealignment_help'] = 'Resets the toggle alignment to follow the site default value.';
-$string['resetalltogglealignment_help'] = 'Resets the toggle alignment to follow the site default value.';
+$string['resetalltogglealignment_help'] = 'Resets all the toggle alignment to follow the site default value.';
 
 // Icon position - CONTRIB-4470.
 $string['settoggleiconposition'] = 'Icon position';
@@ -188,6 +229,10 @@ $string['defaulttoggleiconposition_desc'] = 'States if the icon should be on the
 // Icon set enhancement.
 $string['settoggleiconset'] = 'Icon set';
 $string['settoggleiconset_help'] = 'Sets the icon set of the toggle.';
+$string['settoggleiconfontclosed'] = 'Closed toggle icon font';
+$string['settoggleiconfontclosed_help'] = 'When \'toggleiconset\' is set to \'Icon font\', this states the default CSS classes to use for the closed icon, i.e. see the FontAwesome icon classes.  If set to \'-\' then the default is used.';
+$string['settoggleiconfontopen'] = 'Open toggle icon font';
+$string['settoggleiconfontopen_help'] = 'When \'toggleiconset\' is set to \'Icon font\', this states the default CSS classes to use for the open icon, i.e. see the FontAwesome icon classes.  If set to \'-\' then the default is used.';
 $string['settoggleallhover'] = 'Toggle all icon hover';
 $string['settoggleallhover_help'] = 'Sets if the toggle all icons will change when the mouse moves over them.';
 $string['arrow'] = 'Arrow';
@@ -204,6 +249,7 @@ $string['smiley'] = 'Smiley';
 $string['square'] = 'Square';
 $string['sunmoon'] = 'Sun / Moon';
 $string['switch'] = 'Switch';
+$string['tif'] = 'Icon font';
 $string['resettoggleiconset'] = 'Toggle icon set';
 $string['resetalltoggleiconset'] = 'Toggle icon sets';
 $string['resettoggleiconset_help'] = 'Resets the toggle icon set and toggle all hover to follow the site default value.';
@@ -214,12 +260,14 @@ $string['onesection'] = 'One section';
 $string['onesection_help'] = 'States if only one section should be open at any given time.  Note: Ignored when editing to allow activities and resources to be moved around the sections.';
 $string['defaultonesection'] = 'One section';
 $string['defaultonesection_desc'] = "States if only one section should be open at any given time.  Note: Ignored when editing to allow activities and resources to be moved around the sections.";
+$string['defaultonesectioniconfont'] = 'One section icon font';
+$string['defaultonesectioniconfont_desc'] = 'State the icon font class to use for the one section link icon, i.e. see the FontAwesome icon classes.  If empty, then default icon \'one_section\' in the format\'s \'pix\' folder will be used.';
 
 // Site Administration -> Plugins -> Course formats -> Collapsed Topics.
 $string['defaultheadingsub'] = 'Defaults';
-$string['defaultheadingsubdesc'] = 'Default settings';
+$string['defaultheadingsubdesc'] = 'Default settings that can be overridden at the course level';
 $string['configurationheadingsub'] = 'Configuration';
-$string['configurationheadingsubdesc'] = 'Configuration settings';
+$string['configurationheadingsubdesc'] = 'Site level configuration settings';
 
 $string['off'] = 'Off';
 $string['on'] = 'On';
@@ -239,7 +287,7 @@ $string['defaultlayoutcolumns'] = 'Number of columns';
 $string['defaultlayoutcolumns_desc'] = "Number of columns between one and four.";
 
 $string['defaultlayoutcolumnorientation'] = 'Column orientation';
-$string['defaultlayoutcolumnorientation_desc'] = "The default column orientation: Vertical or Horizontal.";
+$string['defaultlayoutcolumnorientation_desc'] = 'The default column orientation: Dynamic - Number sections per \'row\' adjust to window size, \'Column\' setting not currently used.<br>Horizontal - Sections go left to right.<br>Vertical - Sections go top to bottom.';
 
 $string['defaulttgfgcolour'] = 'Toggle foreground colour';
 $string['defaulttgfgcolour_desc'] = "Toggle foreground colour in hexidecimal RGB.";
@@ -257,7 +305,13 @@ $string['defaulttogglealignment'] = 'Toggle text alignment';
 $string['defaulttogglealignment_desc'] = "'Left', 'Centre' or 'Right'.";
 
 $string['defaulttoggleiconset'] = 'Toggle icon set';
-$string['defaulttoggleiconset_desc'] = '<table><tbody><tr><td>Arrow</td><td>{$a->arrow}</td></tr><tr><td>Bulb</td><td>{$a->bulb}</td></tr><tr><td>Cloud</td><td>{$a->cloud}</td></tr><tr><td>Eye</td><td>{$a->eye}</td></tr><tr><td>Folder</td><td>{$a->folder}</td></tr><tr><td>Ground Signal</td><td>{$a->groundsignal}</td></tr><tr><td>Light Emitting Diode</td><td>{$a->led}</td></tr><tr><td>Point</td><td>{$a->point}</td></tr><tr><td>Power</td><td>{$a->power}</td></tr><tr><td>Radio</td><td>{$a->radio}</td></tr><tr><td>Smiley</td><td>{$a->smiley}</td></tr><tr><td>Square</td><td>{$a->square}</td></tr><tr><td>Sun / Moon</td><td>{$a->sunmoon}</td></tr><tr><td>Switch</td><td>{$a->switch}</td></tr></tbody></table>';
+$string['defaulttoggleiconset_desc'] = '<table><tbody><tr><td>Arrow</td><td>{$a->arrow}</td></tr><tr><td>Bulb</td><td>{$a->bulb}</td></tr><tr><td>Cloud</td><td>{$a->cloud}</td></tr><tr><td>Eye</td><td>{$a->eye}</td></tr><tr><td>Folder</td><td>{$a->folder}</td></tr><tr><td>Ground Signal</td><td>{$a->groundsignal}</td></tr><tr><td>Light Emitting Diode</td><td>{$a->led}</td></tr><tr><td>Point</td><td>{$a->point}</td></tr><tr><td>Power</td><td>{$a->power}</td></tr><tr><td>Radio</td><td>{$a->radio}</td></tr><tr><td>Smiley</td><td>{$a->smiley}</td></tr><tr><td>Square</td><td>{$a->square}</td></tr><tr><td>Sun / Moon</td><td>{$a->sunmoon}</td></tr><tr><td>Switch</td><td>{$a->switch}</td></tr><tr><td>Icon font</td><td>{$a->tif}</td></tr></tbody></table>';
+
+$string['defaulttoggleiconfontclosed'] = 'Closed toggle icon font';
+$string['defaulttoggleiconfontclosed_desc'] = 'When \'defaulttoggleiconset\' is set to \'Icon font\', this states the default CSS classes to use for the closed icon, i.e. see the FontAwesome icon classes.';
+
+$string['defaulttoggleiconfontopen'] = 'Open toggle icon font';
+$string['defaulttoggleiconfontopen_desc'] = 'When \'defaulttoggleiconset\' is set to \'Icon font\', this states the default CSS classes to use for the open icon, i.e. see the FontAwesome icon classes.';
 
 $string['defaulttoggleallhover'] = 'Toggle all icon hovers';
 $string['defaulttoggleallhover_desc'] = "'No' or 'Yes'.";
@@ -291,7 +345,7 @@ $string['defaulttgbghvropacity_desc'] = "Toggle background on hover opacity betw
 
 // Toggle icon size.
 $string['defaulttoggleiconsize'] = 'Toggle icon size';
-$string['defaulttoggleiconsize_desc'] = "Icon size: Small = 16px, Medium = 24px and Large = 32px.";
+$string['defaulttoggleiconsize_desc'] = "Icon size: Small = 16px, Medium = 24px and Large = 32px, or Icon font: Small = 0.8em, Medium = 1.2em and Large = 1.8em.";
 $string['small'] = 'Small';
 $string['medium'] = 'Medium';
 $string['large'] = 'Large';
@@ -365,6 +419,7 @@ $string['topcoll:changelayout'] = 'Change or reset the layout';
 $string['topcoll:changecolour'] = 'Change or reset the colour';
 $string['topcoll:changetogglealignment'] = 'Change or reset the toggle alignment';
 $string['topcoll:changetoggleiconset'] = 'Change or reset the toggle icon set';
+$string['topcoll:changeactivitymeta'] = 'Change or reset the activity meta';
 
 // Instructions.
 $string['instructions'] = 'Instructions: Clicking on the section name will show / hide the section.';
@@ -378,44 +433,18 @@ $string['resetdisplayinstructions_help'] = 'Resets the display instructions to f
 $string['resetalldisplayinstructions_help'] = 'Resets the display instructions to follow the site default value.';
 
 // Activity display *********************************.
-$string['answered'] = 'Answered';
-$string['attempted'] = 'Attempted';
-$string['contributed'] = 'Contributed';
-$string['draft'] = 'Not published to students';
-$string['due'] = 'Due {$a}';
-$string['expired'] = 'Expired';
 $string['feedbackavailable'] = 'Feedback available';
-$string['notanswered'] = 'Not answered';
-$string['notattempted'] = 'Not attempted';
-$string['notcontributed'] = 'Not contributed';
-$string['notopen'] = 'Not open';
-$string['notsubmitted'] = 'Not submitted';
-$string['overdue'] = 'Overdue';
-$string['reopened'] = 'Reopened';
-$string['submitted'] = 'Submitted';
-
 $string['xofyanswered'] = '{$a->completed} of {$a->participants} answered';
 $string['xofyattempted'] = '{$a->completed} of {$a->participants} attempted';
 $string['xofycontributed'] = '{$a->completed} of {$a->participants} contributed';
+$string['xofyposted'] = '{$a->completed} of {$a->participants} posted';
 $string['xofysubmitted'] = '{$a->completed} of {$a->participants} submitted';
+$string['xanswered'] = '{$a->completed} answered';
+$string['xattempted'] = '{$a->completed} attempted';
+$string['xcontributed'] = '{$a->completed} contributed';
+$string['xposted'] = '{$a->completed} posted';
+$string['xsubmitted'] = '{$a->completed} submitted';
 $string['xungraded'] = '{$a} ungraded';
-
-$string['checked'] = 'Checked';
-$string['warning'] = 'Warning';
-
-$string['coursesectionactivityfurtherinformation'] = 'Course page further information';
-$string['coursesectionactivityfurtherinformationassign'] = 'Show assignment information';
-$string['coursesectionactivityfurtherinformationassigndesc'] = 'Show assignment information, such as due date, submission status.  For teachers / admins, show number of submissions.';
-$string['coursesectionactivityfurtherinformationquiz'] = 'Show quiz information';
-$string['coursesectionactivityfurtherinformationquizdesc'] = 'Show quiz information, such as submission status.  For teachers / admins, show number of submissions.';
-$string['coursesectionactivityfurtherinformationchoice'] = 'Show choice information';
-$string['coursesectionactivityfurtherinformationchoicedesc'] = 'Show choice information, such as submission status.  For teachers / admins, show number of submissions.';
-$string['coursesectionactivityfurtherinformationfeedback'] = 'Show feedback information';
-$string['coursesectionactivityfurtherinformationfeedbackdesc'] = 'Show feedback information, such as submission status.  For teachers / admins, show number of submissions.';
-$string['coursesectionactivityfurtherinformationlesson'] = 'Show lesson information';
-$string['coursesectionactivityfurtherinformationlessondesc'] = 'Show lesson information, such as submission status.  For teachers / admins, show number of submissions.';
-$string['coursesectionactivityfurtherinformationdata'] = 'Show database information';
-$string['coursesectionactivityfurtherinformationdatadesc'] = 'Show data information, such as submission status.  For teachers / admins, show number of submissions.';
 
 // Duplicate section.
 $string['creatingsection'] = 'Creating new section';
@@ -427,10 +456,31 @@ $string['progresscounter'] = 'Duplicating activities ({$a->current}/{$a->size})'
 $string['progressfull'] = 'Duplicating topic';
 $string['rebuildcoursecache'] = 'Rebuild course cache';
 
-// Privacy.
-$string['privacy:metadata:preference:toggle'] = 'The state of the toggles on a course.';
-$string['privacy:request:preference:toggle'] = 'The course id "{$a->name}" has the value "{$a->value}" which represents "{$a->decoded}" for the state of the toggles.';
-
 // Readme.
 $string['readme_title'] = 'Collapsed Topics read-me';
 $string['readme_desc'] = 'Please click on \'{$a->url}\' for lots more information about Collapsed Topics.';
+
+// Toggle Display Blocks.
+$string['defaultdisplayblocks'] = 'Blocks to display';
+$string['defaultdisplayblocks_desc'] = "Choose the blocks to display in the course when it is first created and this format is selected.  Use the 'Ctrl' key in combination with the mouse to select more than one or none.  Note: This setting will only apply at actual course creation and no other time, i.e. changing to Collapsed Topics from another format.";
+$string['defaultdisplayblocksloc'] = 'Block location for display';
+$string['defaultdisplayblocksloc_desc'] = "Choose the location for the blocks chosen above to display, pre or post side.";
+$string['sidepost'] = 'Post';
+$string['sidepre'] = 'Pre';
+
+// Information.
+$string['information'] = 'Information';
+$string['informationsettings'] = 'Information settings';
+$string['informationsettingsdesc'] = 'Collapsed Topics format information';
+$string['settings'] = 'Settings';
+$string['settingssettings'] = 'Settings settings';
+$string['settingssettingsdesc'] = 'Collapsed Topics format settings';
+$string['versioninfo'] = 'Release {$a->release}, version {$a->version} on Moodle {$a->moodle}';
+$string['versionalpha'] = 'Alpha version - Almost certainly contains bugs.  This is a development version for developers \'only\'!  Don\'t even think of installing on a production server!';
+$string['versionbeta'] = 'Beta version - Likely to contain bugs.  Ready for testing by administrators on a test server only.';
+$string['versionrc'] = 'Release candidate version - May contain bugs.  Check completely on a test server before considering on a production server.';
+$string['versionstable'] = 'Stable version - Could contain bugs.  Check on a test server before installing on your production server.';
+
+// Privacy.
+$string['privacy:metadata:preference:toggle'] = 'The state of the toggles on a course.';
+$string['privacy:request:preference:toggle'] = 'The course id "{$a->name}" has the value "{$a->value}" which represents "{$a->decoded}" for the state of the toggles.';