+++ course/format/topcoll/settings.php
@@ -32,8 +32,30 @@
  */
 defined('MOODLE_INTERNAL') || die;
 
+$settings = null;
+$ADMIN->add('formatsettings', new admin_category('format_topcoll', get_string('pluginname', 'format_topcoll')));
+
+// Information.
+$page = new admin_settingpage('format_topcoll_information',
+    get_string('information', 'format_topcoll'));
+
 if ($ADMIN->fulltree) {
-    $settings->add(new admin_setting_heading('format_topcoll_defaults',
+    $page->add(new admin_setting_heading('format_topcoll_information', '',
+        format_text(get_string('informationsettingsdesc', 'format_topcoll'), FORMAT_MARKDOWN)));
+
+    // Information.
+    $page->add(new \format_topcoll\admin_setting_information('format_topcoll/formatinformation', '', '', 400));
+
+    // Support.md.
+    $page->add(new \format_topcoll\admin_setting_markdown('format_topcoll/formatsupport', '', '', 'Support.md'));
+}
+$ADMIN->add('format_topcoll', $page);
+
+// Settings.
+$page = new admin_settingpage('format_topcoll_settings',
+    get_string('settings', 'format_topcoll'));
+if ($ADMIN->fulltree) {
+    $page->add(new admin_setting_heading('format_topcoll_defaults',
             get_string('defaultheadingsub', 'format_topcoll'),
             format_text(get_string('defaultheadingsubdesc', 'format_topcoll'), FORMAT_MARKDOWN)));
 
@@ -46,20 +68,51 @@ if ($ADMIN->fulltree) {
         1 => new lang_string('no'), // No.
         2 => new lang_string('yes')   // Yes.
     );
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+
+    /* Toggle display block choices */
+    $name = 'format_topcoll/defaultdisplayblocks';
+    $title = get_string('defaultdisplayblocks', 'format_topcoll');
+    $description = get_string('defaultdisplayblocks_desc', 'format_topcoll');
+    $choices = core_plugin_manager::instance()->get_enabled_plugins('block');
+    // Change the value of the array to have the real string defined in the language file.
+    foreach ($choices as $key => $blockname) {
+        $choices[$key] = get_string('pluginname', 'block_' . $key);
+    }
+    /* See if our desired default blocks '$defaultsearchlist' are in the list of available
+       blocks '$choices' created above, and if so - add each of them to the '$default' array for use. */
+    $default = array();
+    $defaultsearchlist = array('search_forums', 'news_items', 'calendar_upcoming', 'recent_activity');
+    foreach ($defaultsearchlist as $defaultblk) {
+        if (array_key_exists($defaultblk, $choices)) {
+            array_push($default, $defaultblk);
+        }
+    }
+    $page->add(new admin_setting_configmultiselect($name, $title, $description, $default, $choices));
+
+    // Toggle blocks location. 1 = pre, 2 = post.
+    $name = 'format_topcoll/defaultdisplayblocksloc';
+    $title = get_string('defaultdisplayblocksloc', 'format_topcoll');
+    $description = get_string('defaultdisplayblocksloc_desc', 'format_topcoll');
+    $default = 1;
+    $choices = array(
+        1 => new lang_string('sidepre', 'format_topcoll'),   // Pre.
+        2 => new lang_string('sidepost', 'format_topcoll'),  // Post.
+    );
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
 
     /* Layout configuration.
-      Here you can see what numbers in the array represent what layout for setting the default value below.
-      1 => Toggle word, toggle section x and section number - default.
-      2 => Toggle word and section number.
-      3 => Toggle word and toggle section x.
-      4 => Toggle word.
-      5 => Toggle section x and section number.
-      6 => Section number.
-      7 => No additions.
-      8 => Toggle section x.
-      Default layout to use - used when a new Collapsed Topics course is created or an old one is accessed for the first time
-      after installing this functionality introduced in CONTRIB-3378. */
+       Here you can see what numbers in the array represent what layout for setting the default value below.
+       1 => Toggle word, toggle section x and section number - default.
+       2 => Toggle word and section number.
+       3 => Toggle word and toggle section x.
+       4 => Toggle word.
+       5 => Toggle section x and section number.
+       6 => Section number.
+       7 => No additions.
+       8 => Toggle section x.
+       Default layout to use - used when a new Collapsed Topics course is created or an old one is accessed for the first time
+       after installing this functionality introduced in CONTRIB-3378. */
     $name = 'format_topcoll/defaultlayoutelement';
     $title = get_string('defaultlayoutelement', 'format_topcoll');
     $description = get_string('defaultlayoutelement_descpositive', 'format_topcoll');
@@ -74,17 +127,17 @@ if ($ADMIN->fulltree) {
         6 => new lang_string('setlayout_section_number', 'format_topcoll'), // Section number.
         7 => new lang_string('setlayout_no_additions', 'format_topcoll')                     // No additions.
     );
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
 
     /* Structure configuration.
-      Here so you can see what numbers in the array represent what structure for setting the default value below.
-      1 => Topic.
-      2 => Week.
-      3 => Latest Week First.
-      4 => Current Topic First.
-      5 => Day.
-      Default structure to use - used when a new Collapsed Topics course is created or an old one is accessed for the first time
-      after installing this functionality introduced in CONTRIB-3378. */
+       Here so you can see what numbers in the array represent what structure for setting the default value below.
+       1 => Topic.
+       2 => Week.
+       3 => Latest Week First.
+       4 => Current Topic First.
+       5 => Day.
+       Default structure to use - used when a new Collapsed Topics course is created or an old one is accessed for the first time
+       after installing this functionality introduced in CONTRIB-3378. */
     $name = 'format_topcoll/defaultlayoutstructure';
     $title = get_string('defaultlayoutstructure', 'format_topcoll');
     $description = get_string('defaultlayoutstructure_desc', 'format_topcoll');
@@ -96,7 +149,19 @@ if ($ADMIN->fulltree) {
         4 => new lang_string('setlayoutstructurecurrenttopicfirst', 'format_topcoll'), // Current Topic First.
         5 => new lang_string('setlayoutstructureday', 'format_topcoll')                // Day.
     );
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+
+    // Default column orientation - 1 = vertical and 2 = horizontal.
+    $name = 'format_topcoll/defaultlayoutcolumnorientation';
+    $title = get_string('defaultlayoutcolumnorientation', 'format_topcoll');
+    $description = get_string('defaultlayoutcolumnorientation_desc', 'format_topcoll');
+    $default = 3;
+    $choices = array(
+        3 => new lang_string('columndynamic', 'format_topcoll'),
+        2 => new lang_string('columnhorizontal', 'format_topcoll'),
+        1 => new lang_string('columnvertical', 'format_topcoll')
+    );
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
 
     // Default number of columns between 1 and 4.
     $name = 'format_topcoll/defaultlayoutcolumns';
@@ -109,18 +174,7 @@ if ($ADMIN->fulltree) {
         3 => new lang_string('three', 'format_topcoll'), // Three.
         4 => new lang_string('four', 'format_topcoll')   // Four.
     );
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
-
-    // Default column orientation - 1 = vertical and 2 = horizontal.
-    $name = 'format_topcoll/defaultlayoutcolumnorientation';
-    $title = get_string('defaultlayoutcolumnorientation', 'format_topcoll');
-    $description = get_string('defaultlayoutcolumnorientation_desc', 'format_topcoll');
-    $default = 2;
-    $choices = array(
-        1 => new lang_string('columnvertical', 'format_topcoll'),
-        2 => new lang_string('columnhorizontal', 'format_topcoll') // Default.
-    );
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
 
     /* Toggle all enabled - 1 = no, 2 = yes. */
     $name = 'format_topcoll/defaulttoggleallenabled';
@@ -131,7 +185,7 @@ if ($ADMIN->fulltree) {
         1 => new lang_string('no'), // No.
         2 => new lang_string('yes') // Yes.
     );
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
 
     /* View single section enabled - 1 = no, 2 = yes. */
     $name = 'format_topcoll/defaultviewsinglesectionenabled';
@@ -142,7 +196,7 @@ if ($ADMIN->fulltree) {
         1 => new lang_string('no'), // No.
         2 => new lang_string('yes') // Yes.
     );
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
 
     // Toggle text alignment.
     // 1 = left, 2 = center and 3 = right - done this way to avoid typos.
@@ -155,7 +209,7 @@ if ($ADMIN->fulltree) {
         2 => new lang_string('center', 'format_topcoll'), // Centre.
         3 => new lang_string('right', 'format_topcoll')   // Right.
     );
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
 
     // Toggle icon position.
     // 1 = left and 2 = right - done this way to avoid typos.
@@ -167,23 +221,25 @@ if ($ADMIN->fulltree) {
         1 => new lang_string('left', 'format_topcoll'), // Left.
         2 => new lang_string('right', 'format_topcoll')   // Right.
     );
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
 
-    // Toggle icon set.
-    // arrow        => Arrow icon set.
-    // bulb         => Bulb icon set.
-    // cloud        => Cloud icon set.
-    // eye          => Eye icon set.
-    // folder       => Folder icon set.
-    // groundsignal => Ground signal set.
-    // led          => LED icon set.
-    // point        => Point icon set.
-    // power        => Power icon set.
-    // radio        => Radio icon set.
-    // smiley       => Smiley icon set.
-    // square       => Square icon set.
-    // sunmoon      => Sun / Moon icon set.
-    // switch       => Switch icon set.
+    /* Toggle icon set.
+       arrow        => Arrow icon set.
+       bulb         => Bulb icon set.
+       cloud        => Cloud icon set.
+       eye          => Eye icon set.
+       folder       => Folder icon set.
+       groundsignal => Ground signal set.
+       led          => LED icon set.
+       point        => Point icon set.
+       power        => Power icon set.
+       radio        => Radio icon set.
+       smiley       => Smiley icon set.
+       square       => Square icon set.
+       sunmoon      => Sun / Moon icon set.
+       switch       => Switch icon set.
+       tif          => Icon font.
+    */
     $iconseticons = array(
         'arrow' => $OUTPUT->pix_icon('arrow_right', get_string('arrow', 'format_topcoll'), 'format_topcoll'),
         'bulb' => $OUTPUT->pix_icon('bulb_off', get_string('bulb', 'format_topcoll'), 'format_topcoll'),
@@ -198,12 +254,13 @@ if ($ADMIN->fulltree) {
         'smiley' => $OUTPUT->pix_icon('smiley_on', get_string('smiley', 'format_topcoll'), 'format_topcoll'),
         'square' => $OUTPUT->pix_icon('square_on', get_string('square', 'format_topcoll'), 'format_topcoll'),
         'sunmoon' => $OUTPUT->pix_icon('sunmoon_on', get_string('sunmoon', 'format_topcoll'), 'format_topcoll'),
-        'switch' => $OUTPUT->pix_icon('switch_on', get_string('switch', 'format_topcoll'), 'format_topcoll')
+        'switch' => $OUTPUT->pix_icon('switch_on', get_string('switch', 'format_topcoll'), 'format_topcoll'),
+        'tif' => $OUTPUT->pix_icon('iconfont', get_string('tif', 'format_topcoll'), 'format_topcoll')
     );
     $name = 'format_topcoll/defaulttoggleiconset';
     $title = get_string('defaulttoggleiconset', 'format_topcoll');
     $description = get_string('defaulttoggleiconset_desc', 'format_topcoll', $iconseticons);
-    $default = 'arrow';
+    $default = 'tif';
     $choices = array(
         'arrow' => new lang_string('arrow', 'format_topcoll'), // Arrow icon set.
         'bulb' => new lang_string('bulb', 'format_topcoll'), // Bulb icon set.
@@ -218,9 +275,24 @@ if ($ADMIN->fulltree) {
         'smiley' => new lang_string('smiley', 'format_topcoll'), // Smiley icon set.
         'square' => new lang_string('square', 'format_topcoll'), // Square icon set.
         'sunmoon' => new lang_string('sunmoon', 'format_topcoll'), // Sun / Moon icon set.
-        'switch' => new lang_string('switch', 'format_topcoll') // Switch icon set.
+        'switch' => new lang_string('switch', 'format_topcoll'), // Switch icon set.
+        'tif' => new lang_string('tif', 'format_topcoll') // Toggle icon font.
     );
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+
+    $name = 'format_topcoll/defaulttoggleiconfontclosed';
+    $title = get_string('defaulttoggleiconfontclosed', 'format_topcoll');
+    $description = get_string('defaulttoggleiconfontclosed_desc', 'format_topcoll');
+    $default = 'fa fa-chevron-circle-right';
+    $setting = new admin_setting_configtext($name, $title, $description, $default, PARAM_TEXT);
+    $page->add($setting);
+
+    $name = 'format_topcoll/defaulttoggleiconfontopen';
+    $title = get_string('defaulttoggleiconfontopen', 'format_topcoll');
+    $description = get_string('defaulttoggleiconfontopen_desc', 'format_topcoll');
+    $default = 'fa fa-chevron-circle-down';
+    $setting = new admin_setting_configtext($name, $title, $description, $default, PARAM_TEXT);
+    $page->add($setting);
 
     /* One section - 1 = no, 2 = yes. */
     $name = 'format_topcoll/defaultonesection';
@@ -231,11 +303,19 @@ if ($ADMIN->fulltree) {
         1 => new lang_string('no'), // No.
         2 => new lang_string('yes') // Yes.
     );
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
 
-    // Toggle all icon hovers.
-    // 1 => No.
-    // 2 => Yes.
+    /* One section icon font */
+    $name = 'format_topcoll/defaultonesectioniconfont';
+    $title = get_string('defaultonesectioniconfont', 'format_topcoll');
+    $description = get_string('defaultonesectioniconfont_desc', 'format_topcoll');
+    $default = 'fa fa-dot-circle-o';
+    $setting = new admin_setting_configtext($name, $title, $description, $default, PARAM_TEXT);
+    $page->add($setting);
+
+    /* Toggle all icon hovers.
+       1 => No.
+       2 => Yes. */
     $name = 'format_topcoll/defaulttoggleallhover';
     $title = get_string('defaulttoggleallhover', 'format_topcoll');
     $description = get_string('defaulttoggleallhover_desc', 'format_topcoll');
@@ -244,7 +324,7 @@ if ($ADMIN->fulltree) {
         1 => new lang_string('no'),
         2 => new lang_string('yes')
     );
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
 
     $opacityvalues = array(
         '0.0' => '0.0',
@@ -264,65 +344,65 @@ if ($ADMIN->fulltree) {
     $name = 'format_topcoll/defaulttoggleforegroundcolour';
     $title = get_string('defaulttgfgcolour', 'format_topcoll');
     $description = get_string('defaulttgfgcolour_desc', 'format_topcoll');
-    $default = '#000000';
+    $default = '#eeeeee';
     $setting = new admin_setting_configcolourpicker($name, $title, $description, $default);
-    $settings->add($setting);
+    $page->add($setting);
 
     // Default toggle foreground opacity between 0 and 1 in 0.1 increments.
     $name = 'format_topcoll/defaulttoggleforegroundopacity';
     $title = get_string('defaulttgfgopacity', 'format_topcoll');
     $description = get_string('defaulttgfgopacity_desc', 'format_topcoll');
     $default = '1.0';
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $opacityvalues));
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $opacityvalues));
 
     // Default toggle foreground hover colour in hexadecimal RGB with preceding '#'.
     $name = 'format_topcoll/defaulttoggleforegroundhovercolour';
     $title = get_string('defaulttgfghvrcolour', 'format_topcoll');
     $description = get_string('defaulttgfghvrcolour_desc', 'format_topcoll');
-    $default = '#888888';
+    $default = '#ffffff';
     $setting = new admin_setting_configcolourpicker($name, $title, $description, $default);
-    $settings->add($setting);
+    $page->add($setting);
 
     // Default toggle foreground hover opacity between 0 and 1 in 0.1 increments.
     $name = 'format_topcoll/defaulttoggleforegroundhoveropacity';
     $title = get_string('defaulttgfghvropacity', 'format_topcoll');
     $description = get_string('defaulttgfghvropacity_desc', 'format_topcoll');
     $default = '1.0';
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $opacityvalues));
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $opacityvalues));
 
     // Default toggle background colour in hexadecimal RGB with preceding '#'.
     $name = 'format_topcoll/defaulttogglebackgroundcolour';
     $title = get_string('defaulttgbgcolour', 'format_topcoll');
     $description = get_string('defaulttgbgcolour_desc', 'format_topcoll');
-    $default = '#e2e2f2';
+    $default = '#1177d1';
     $setting = new admin_setting_configcolourpicker($name, $title, $description, $default);
-    $settings->add($setting);
+    $page->add($setting);
 
     // Default toggle background opacity between 0 and 1 in 0.1 increments.
     $name = 'format_topcoll/defaulttogglebackgroundopacity';
     $title = get_string('defaulttgbgopacity', 'format_topcoll');
     $description = get_string('defaulttgbgopacity_desc', 'format_topcoll');
     $default = '1.0';
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $opacityvalues));
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $opacityvalues));
 
     // Default toggle background hover colour in hexadecimal RGB with preceding '#'.
     $name = 'format_topcoll/defaulttogglebackgroundhovercolour';
     $title = get_string('defaulttgbghvrcolour', 'format_topcoll');
     $description = get_string('defaulttgbghvrcolour_desc', 'format_topcoll');
-    $default = '#eeeeff';
+    $default = '#1482E2';
     $setting = new admin_setting_configcolourpicker($name, $title, $description, $default);
-    $settings->add($setting);
+    $page->add($setting);
 
     // Default toggle background hover opacity between 0 and 1 in 0.1 increments.
     $name = 'format_topcoll/defaulttogglebackgroundhoveropacity';
     $title = get_string('defaulttgbghvropacity', 'format_topcoll');
     $description = get_string('defaulttgbghvropacity_desc', 'format_topcoll');
     $default = '1.0';
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $opacityvalues));
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $opacityvalues));
 
-    // Show the section summary when collapsed.
-    // 1 => No.
-    // 2 => Yes.
+    /* Show the section summary when collapsed.
+       1 => No.
+       2 => Yes. */
     $name = 'format_topcoll/defaultshowsectionsummary';
     $title = get_string('defaultshowsectionsummary', 'format_topcoll');
     $description = get_string('defaultshowsectionsummary_desc', 'format_topcoll');
@@ -331,9 +411,9 @@ if ($ADMIN->fulltree) {
         1 => new lang_string('no'),
         2 => new lang_string('yes')
     );
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
 
-    $settings->add(new admin_setting_heading('format_topcoll_configuration',
+    $page->add(new admin_setting_heading('format_topcoll_configuration',
             get_string('configurationheadingsub', 'format_topcoll'),
             format_text(get_string('configurationheadingsubdesc', 'format_topcoll'), FORMAT_MARKDOWN)));
 
@@ -349,11 +429,11 @@ if ($ADMIN->fulltree) {
         0 => new lang_string('off', 'format_topcoll'), // Off.
         1 => new lang_string('on', 'format_topcoll')   // On.
     );
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
 
-    // Toggle preference for the first time a user accesses a course.
-    // 0 => All closed.
-    // 1 => All open.
+    /* Toggle preference for the first time a user accesses a course.
+       0 => All closed.
+       1 => All open. */
     $name = 'format_topcoll/defaultuserpreference';
     $title = get_string('defaultuserpreference', 'format_topcoll');
     $description = get_string('defaultuserpreference_desc', 'format_topcoll');
@@ -362,7 +442,7 @@ if ($ADMIN->fulltree) {
         0 => new lang_string('topcollclosed', 'format_topcoll'),
         1 => new lang_string('topcollopened', 'format_topcoll')
     );
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
 
     // Toggle icon size.
     $name = 'format_topcoll/defaulttoggleiconsize';
@@ -374,13 +454,13 @@ if ($ADMIN->fulltree) {
         'tc-medium' => new lang_string('medium', 'format_topcoll'),
         'tc-large' => new lang_string('large', 'format_topcoll')
     );
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
 
     // Toggle border radius top left.
     $name = 'format_topcoll/defaulttoggleborderradiustl';
     $title = get_string('defaulttoggleborderradiustl', 'format_topcoll');
     $description = get_string('defaulttoggleborderradiustl_desc', 'format_topcoll');
-    $default = '0.7';
+    $default = '0.0';
     $choices = array(
         '0.0' => new lang_string('em0_0', 'format_topcoll'),
         '0.1' => new lang_string('em0_1', 'format_topcoll'),
@@ -424,28 +504,28 @@ if ($ADMIN->fulltree) {
         '3.9' => new lang_string('em3_9', 'format_topcoll'),
         '4.0' => new lang_string('em4_0', 'format_topcoll')
     );
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
 
     // Toggle border radius top right.
     $name = 'format_topcoll/defaulttoggleborderradiustr';
     $title = get_string('defaulttoggleborderradiustr', 'format_topcoll');
     $description = get_string('defaulttoggleborderradiustr_desc', 'format_topcoll');
-    $default = '0.7';
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+    $default = '1.2';
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
 
     // Toggle border radius bottom right.
     $name = 'format_topcoll/defaulttoggleborderradiusbr';
     $title = get_string('defaulttoggleborderradiusbr', 'format_topcoll');
     $description = get_string('defaulttoggleborderradiusbr_desc', 'format_topcoll');
-    $default = '0.7';
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+    $default = '0.4';
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
 
     // Toggle border radius bottom left.
     $name = 'format_topcoll/defaulttoggleborderradiusbl';
     $title = get_string('defaulttoggleborderradiusbl', 'format_topcoll');
     $description = get_string('defaulttoggleborderradiusbl_desc', 'format_topcoll');
-    $default = '0.7';
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+    $default = '0.2';
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
 
     /* Format responsive.  Turn on to support a non responsive theme theme. */
     $name = 'format_topcoll/formatresponsive';
@@ -456,11 +536,11 @@ if ($ADMIN->fulltree) {
         0 => new lang_string('off', 'format_topcoll'), // Off.
         1 => new lang_string('on', 'format_topcoll')   // On.
     );
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
 
-    // Show the section summary when collapsed.
-    // 1 => No.
-    // 2 => Yes.
+    /* Show the section summary when collapsed.
+       1 => No.
+       2 => Yes. */
     $name = 'format_topcoll/defaultshowsectionsummary';
     $title = get_string('defaultshowsectionsummary', 'format_topcoll');
     $description = get_string('defaultshowsectionsummary_desc', 'format_topcoll');
@@ -469,13 +549,34 @@ if ($ADMIN->fulltree) {
         1 => new lang_string('no'),
         2 => new lang_string('yes')
     );
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
 
     // Course Activity Further Information section heading.
     $name = 'format_topcoll/coursesectionactivityfurtherinformation';
     $heading = get_string('coursesectionactivityfurtherinformation', 'format_topcoll');
-    $setting = new admin_setting_heading($name, $heading, '');
-    $settings->add($setting);
+    $description = get_string('coursesectionactivityfurtherinformation_desc', 'format_topcoll');
+    $setting = new admin_setting_heading($name, $heading, $description);
+    $page->add($setting);
+
+    $name = 'format_topcoll/enableadditionalmoddata';
+    $title = get_string('enableadditionalmoddata', 'format_topcoll');
+    $description = get_string('enableadditionalmoddatadesc', 'format_topcoll');
+    $default = 1;
+    $choices = array(
+        1 => new lang_string('no'),
+        2 => new lang_string('yes')
+    );
+    $setting = new admin_setting_configselect($name, $title, $description, $default, $choices);
+    $setting->set_updatedcallback('cache_helper::purge_all');
+    $page->add($setting);
+
+    $name = 'format_topcoll/courseadditionalmoddatamaxstudents';
+    $title = get_string('courseadditionalmoddatamaxstudents', 'format_topcoll');
+    $description = get_string('courseadditionalmoddatamaxstudentsdesc', 'format_topcoll');
+    $default = 0;
+    $setting = new admin_setting_configtext($name, $title, $description, $default, PARAM_INT);
+    $setting->set_updatedcallback('cache_helper::purge_all');
+    $page->add($setting);
 
     $name = 'format_topcoll/defaultshowadditionalmoddata';
     $title = get_string('defaultshowadditionalmoddata', 'format_topcoll');
@@ -485,7 +586,7 @@ if ($ADMIN->fulltree) {
         1 => new lang_string('no'),
         2 => new lang_string('yes')
     );
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
 
     $name = 'format_topcoll/coursesectionactivityfurtherinformationassign';
     $title = get_string('coursesectionactivityfurtherinformationassign', 'format_topcoll');
@@ -495,7 +596,7 @@ if ($ADMIN->fulltree) {
         1 => new lang_string('no'),
         2 => new lang_string('yes')
     );
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
 
     $name = 'format_topcoll/coursesectionactivityfurtherinformationquiz';
     $title = get_string('coursesectionactivityfurtherinformationquiz', 'format_topcoll');
@@ -505,7 +606,7 @@ if ($ADMIN->fulltree) {
         1 => new lang_string('no'),
         2 => new lang_string('yes')
     );
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
 
     $name = 'format_topcoll/coursesectionactivityfurtherinformationchoice';
     $title = get_string('coursesectionactivityfurtherinformationchoice', 'format_topcoll');
@@ -515,7 +616,7 @@ if ($ADMIN->fulltree) {
         1 => new lang_string('no'),
         2 => new lang_string('yes')
     );
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
 
     $name = 'format_topcoll/coursesectionactivityfurtherinformationfeedback';
     $title = get_string('coursesectionactivityfurtherinformationfeedback', 'format_topcoll');
@@ -525,7 +626,17 @@ if ($ADMIN->fulltree) {
         1 => new lang_string('no'),
         2 => new lang_string('yes')
     );
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+
+    $name = 'format_topcoll/coursesectionactivityfurtherinformationforum';
+    $title = get_string('coursesectionactivityfurtherinformationforum', 'format_topcoll');
+    $description = get_string('coursesectionactivityfurtherinformationforumdesc', 'format_topcoll');
+    $default = 2;
+    $choices = array(
+        1 => new lang_string('no'),
+        2 => new lang_string('yes')
+    );
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
 
     $name = 'format_topcoll/coursesectionactivityfurtherinformationlesson';
     $title = get_string('coursesectionactivityfurtherinformationlesson', 'format_topcoll');
@@ -535,7 +646,7 @@ if ($ADMIN->fulltree) {
         1 => new lang_string('no'),
         2 => new lang_string('yes')
     );
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
 
     $name = 'format_topcoll/coursesectionactivityfurtherinformationdata';
     $title = get_string('coursesectionactivityfurtherinformationdata', 'format_topcoll');
@@ -545,5 +656,6 @@ if ($ADMIN->fulltree) {
         1 => new lang_string('no'),
         2 => new lang_string('yes')
     );
-    $settings->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
+    $page->add(new admin_setting_configselect($name, $title, $description, $default, $choices));
 }
+$ADMIN->add('format_topcoll', $page);